Move the map into it's own component
